--- a/src/diff/mod.rs
+++ b/src/diff/mod.rs
@@ -3,4 +3,5 @@ pub mod dijkstra;
 mod graph;
 pub mod myers_diff;
 pub mod sliders;
+mod stack;
 pub mod unchanged;