@QuarticCat 2022-09-27
Enable PGO could get another >5% speedup. I use this tool to do the optimization https://github.com/Kobzol/cargo-pgo.

```console
$ cargo pgo run -- sample_files/slow_before.rs sample_files/slow_after.rs
$ cargo pgo optimize run -- sample_files/slow_before.rs sample_files/slow_after.rs
```

Before (this time I directly invoke the binary rather than through `cargo run`):

```text
 Performance counter stats for '/home/qc/.cargo/target/release/difft sample_files/slow_before.rs sample_files/slow_after.rs':

            975.13 msec task-clock:u              #    1.000 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             1,423      page-faults:u             #    1.459 K/sec                  
     3,873,705,459      cycles:u                  #    3.973 GHz                    
       312,841,226      stalled-cycles-frontend:u #    8.08% frontend cycles idle   
     1,356,026,289      stalled-cycles-backend:u  #   35.01% backend cycles idle    
     4,416,634,146      instructions:u            #    1.14  insn per cycle         
                                                  #    0.31  stalled cycles per insn
       834,316,552      branches:u                #  855.598 M/sec                  
        18,932,572      branch-misses:u           #    2.27% of all branches        

       0.975598289 seconds time elapsed

       0.947813000 seconds user
       0.026704000 seconds sy
```

After:

```text
 Performance counter stats for '/home/qc/.cargo/target/x86_64-unknown-linux-gnu/release/difft sample_files/slow_before.rs sample_files/slow_after.rs':

            907.64 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             1,441      page-faults:u             #    1.588 K/sec                  
     3,573,701,152      cycles:u                  #    3.937 GHz                    
        81,488,884      stalled-cycles-frontend:u #    2.28% frontend cycles idle   
     1,843,191,214      stalled-cycles-backend:u  #   51.58% backend cycles idle    
     4,175,713,686      instructions:u            #    1.17  insn per cycle         
                                                  #    0.44  stalled cycles per insn
       773,151,834      branches:u                #  851.826 M/sec                  
        18,228,346      branch-misses:u           #    2.36% of all branches        

       0.908512976 seconds time elapsed

       0.860485000 seconds user
       0.046691000 seconds sys
```