@QuarticCat 2022-09-27
Substitute `rpds::Stack` with a simplified one, which removes all unused parts. Nearly 20% speedup.

```
 Performance counter stats for 'cargo run --release -- sample_files/slow_before.rs sample_files/slow_after.rs':

            802.58 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             5,885      page-faults:u             #    7.333 K/sec                  
     3,161,969,385      cycles:u                  #    3.940 GHz                    
        91,496,359      stalled-cycles-frontend:u #    2.89% frontend cycles idle   
     1,254,938,698      stalled-cycles-backend:u  #   39.69% backend cycles idle    
     3,982,788,439      instructions:u            #    1.26  insn per cycle         
                                                  #    0.32  stalled cycles per insn
       752,314,449      branches:u                #  937.376 M/sec                  
        19,227,259      branch-misses:u           #    2.56% of all branches        

       0.803042017 seconds time elapsed

       0.752424000 seconds user
       0.049889000 seconds sys
```