@QuarticCat 2022-09-27
Switching from mimalloc to snmalloc brings a negligible speedup and slightly less memory usage.

Since the time difference is too small, hyperfine is used again.

Before this change:

```text
Benchmark 1: cargo run --release -- sample_files/slow_before.rs sample_files/slow_after.rs
  Time (mean ± σ):     863.1 ms ±  36.6 ms    [User: 816.1 ms, System: 45.9 ms]
  Range (min … max):   824.6 ms … 922.8 ms    10 runs
```

After this change:

```text
Benchmark 1: cargo run --release -- sample_files/slow_before.rs sample_files/slow_after.rs
  Time (mean ± σ):     852.8 ms ±  40.2 ms    [User: 804.4 ms, System: 47.2 ms]
  Range (min … max):   817.9 ms … 929.6 ms    10 runs
```

```text
cargo run --release -- sample_files/slow_before.rs sample_files/slow_after.rs   0.73s  user 0.06s system 99% cpu 0.791 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                464 
page faults from disk:     0
other page faults:         7519
```

If you think such an improvement is worthy then I will commit & push it.