@QuarticCat 2022-09-27
Change a `RefCell` in `Vertex` to `Cell` and save some memory.

Benchmark results (without PGO and snmalloc):

```text
 Performance counter stats for 'cargo run --release -- sample_files/slow_before.rs sample_files/slow_after.rs':

            833.68 msec task-clock:u              #    0.999 CPUs utilized          
                 0      context-switches:u        #    0.000 /sec                   
                 0      cpu-migrations:u          #    0.000 /sec                   
             5,870      page-faults:u             #    7.041 K/sec                  
     3,216,321,688      cycles:u                  #    3.858 GHz                    
       116,517,616      stalled-cycles-frontend:u #    3.62% frontend cycles idle   
     1,287,228,852      stalled-cycles-backend:u  #   40.02% backend cycles idle    
     3,999,693,671      instructions:u            #    1.24  insn per cycle         
                                                  #    0.32  stalled cycles per insn
       753,977,344      branches:u                #  904.398 M/sec                  
        19,331,816      branch-misses:u           #    2.56% of all branches        

       0.834518514 seconds time elapsed

       0.803612000 seconds user
       0.030010000 seconds sys
```

```text
cargo run --release -- sample_files/slow_before.rs sample_files/slow_after.rs   0.76s  user 0.04s system 99% cpu 0.802 total
avg shared (code):         0 KB
avg unshared (data/stack): 0 KB
total (sum):               0 KB
max memory:                463 
page faults from disk:     0
other page faults:         6740
```

I don't know why the number of instructions rised a little bit.